Filter results to only show drinks with all selected ingredients
The ResultsContainer component now filters the results to only show drinks that contain all the selected ingredients. This is done by creating a helper function that counts the number of times each drink appears in the results and then filters the results to only show drinks that appear the same number of times as the number of selected ingredients.