Refactor filtering algorithm
The filtering algorithm was refactored to handle the case where no drinks are common to all results. In this case, the algorithm will display the drinks that are most common to all results. The algorithm was also refactored to use a single object to store the count of each drink, and to use the `Object.entries` method to sort the drinks by their count.