Fix bug in NameSearchResults
The NameSearchResults component was not handling the case where the fetch
was successful but no results were returned. This commit adds a check for
this case and displays a message to the user if no results are found.
Also adds a loading message to be displayed while the fetch is in progress.